rename async.rs in preparation for edition 2018 update
